jawad-livekit/requirements.txt
# Audio Pipeline Dependencies

# LiveKit - Real-time audio/video communication
livekit>=0.11.0
livekit-api>=0.6.0

# WebSockets for ElevenLabs STT
websockets>=12.0

# Audio processing
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# Async utilities
aiohttp>=3.9.0

# LLM Integration - Anthropic Claude for interview evaluation
anthropic>=0.34.0

# Optional: For testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
